+++ rotkehlchen/chain/ethereum/modules/zksync/constants.py
@@ -1,4 +1,5 @@
 from rotkehlchen.chain.evm.types import string_to_evm_address
 
 ZKSYNC_BRIDGE = string_to_evm_address('0xaBEA9132b05A70803a4E85094fD0e1800777fBEF')
+ZKSYNC_LITE_SUNSET_CLAIM = string_to_evm_address('0x0a14B696350546110a0D8acDb86226983af9D2a0')
 CPT_ZKSYNC = 'zksync'